add strategyHelper.tsx
